ai-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
langchain==0.2.6
langchain-community==0.2.6
langchain-ollama==0.1.1
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
pypdf==4.2.0
httpx==0.27.0